+++ src/libraries/functions.h
@@ -17,7 +17,6 @@
 *	Prima di terminare l'esecuzione di un programma, se la funzione ha avuto successo è necessario
 *	liberare la memoria allocata. Ciò può essere fatto tramite la funzione LiberaDatabase o manualmente.
 */
-
 char** CaricaDatabase(const char*, size_t, size_t*);
 
 /* Libera tutta la memoria allocata da CaricaDatabase */
@@ -30,25 +29,25 @@ int CalcolaCompatibilita(char*, char*);
 int CheckInRima(char*, char*);
 
 /* controlla se, dato un verso, ne esiste uno nel database di cui viene fornito un array di stringhe e la sua lunghezza */
-char* CercaRima(char*, char**, int);
+char* CercaRima(char*, char**, size_t);
 
 /* applica la cesura su un verso, se possibile */
 char* ApplicaCesura(char*);
 
 /* cerca in un data base di cui viene fornito un array di stringhe e la sua lunghezza se esiste un verso cesurabile(?) */
-char* CercaCesura(char**, int);
+char* CercaCesura(char**, size_t);
 
 /* controlla se due verrsi sono assonanti */
 int CheckAssonanza(char*, char*);
 
 /* controlla se, dato un verso, ne esiste uno assonante nel database di cui viene forntio un array di stringhe e la sua lunghezza */
-char* CercaAssonanza(char*, char**, int);
+char* CercaAssonanza(char*, char**, size_t);
 
 /* controlla se in un verso è presente una alliterazione */
 int CheckAlliterazione(char*);
 
 /* controlla se, data una lettera e un array di stringhe e la sua lunghezza, se esiste un verso che ha una alliterazione con quella lettera */
-char* CercaAlliterazione(char, char**, int);
+char* CercaAlliterazione(char, char**, size_t);
 
 /* dati due interi calcola la percentuale della lunghezza del minore sul maggiore(il maggiore è il primo dei due) */
 float CalcolaPercentuale(int, int);
